Finished new, grid based, percussion browser.
